feat(shift_plan): change schema a bit, move it out of branch and treat it as stand alone entity
