+++ lib/v0/index.js
@@ -76,5 +76,6 @@ module.exports.product_addon_groups = require('./product_addon_groups')
 module.exports.configurations = require('./configurations')
 module.exports.service_categories = require('./service_categories')
 module.exports.services = require('./services')
+module.exports.shift_plan = require('./shift_plan')
 
 // hygen:injection -  Please, don't delete this line: when running the cli for schema resources the new routes will be automatically added here.
