+++ lib/v0/shift_plan/index.js
@@ -1,0 +1,2 @@
+module.exports.update = require('./update')
+module.exports.read = require('./read')
