+++ lib/v0/shift_plan/update.js
@@ -1,4 +1,4 @@
-const commonResponse = require('../../../common/response')
+const commonResponse = require('../../common/response')
 const base = require('./base')
 
 module.exports.request = {
